+++ python/etl/pipeline.py
@@ -1,0 +1,10 @@
+"""
+Implement commands that interact with AWS Data Pipeline
+"""
+
+
+def show_pipelines(selection):
+    """
+    List the currently installed pipelines, possibly using a subset based on the selection pattern.
+    """
+    pass